backend/pyproject.toml
[project]
name = "farmos-backend"
version = "0.1.0"
description = "FarmOS 2.0 Backend API — IoT sensor data, irrigation control, and farm management"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "chromadb>=1.5.5",
    "defusedxml>=0.7.1",
    "fastapi>=0.135.2",
    "fastmcp>=3.2.0",
    "fpdf2>=2.8.0",
    "httpx>=0.28.0",
    "langchain-core>=1.2.27",
    "langchain-openai>=1.1.12",
    "langchain-upstage>=0.7.0",
    "langgraph>=1.1.6",
    "passlib[bcrypt]>=1.7.4",
    "Pillow>=12.2.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "rapidfuzz>=3.10.0",
    "ruff>=0.6.0",
    "sentence-transformers>=3.3.0",
    "sqlalchemy[asyncio]>=2.0.48",
    "sse-starlette>=3.3.4",
    "tzdata>=2026.1",
    "uvicorn[standard]>=0.42.0",
]

[project.scripts]
# 공익직불 시행지침 PDF 인덱싱 CLI (팀원 신규 셋업 시 자동 실행되지만, 수동 재인덱싱용 편의 명령)
# 사용: cd backend && uv run subsidy-ingest
subsidy-ingest = "app.services.subsidy.gov_rag:run_ingest_pipeline"
# PDF → Markdown 변환만 (Upstage Parse 호출, 최초 1회 또는 PDF 교체 시)
subsidy-parse-pdf = "app.services.subsidy.pdf_ingest:_cli_main"

[tool.uv]
exclude-newer = "7 days"
# [project.scripts] 엔트리포인트 설치를 활성화 (subsidy-ingest, subsidy-parse-pdf 명령어 등록용)
package = true

[tool.setuptools.packages.find]
# data/, chroma_data/, logs/ 같은 런타임 산출물 디렉터리를 패키지 탐색에서 배제
include = ["app*"]
exclude = ["data*", "chroma_data*", "logs*", ".ruff_cache*"]
